Content-chunking/requirements.txt
fastapi
python-pptx
openai==0.28
asyncio
nest_asyncio
python-dotenv
uvicorn
python-dotenv